+++ native-windows-gui/src/controls/combo_box.rs
@@ -3,7 +3,7 @@ use winapi::shared::minwindef::{LPARAM, WPARAM};
 use winapi::um::winuser::{WS_VISIBLE, WS_DISABLED, WS_TABSTOP};
 use crate::win32::base_helper::{check_hwnd, to_utf16, from_utf16};
 use crate::win32::window_helper as wh;
-use crate::{Font, NwgError};
+use crate::{Font, NwgError, VTextAlign, RawEventHandler, unbind_raw_event_handler};
 use super::{ControlHandle, ControlBase};
 use std::cell::{Ref, RefMut, RefCell};
 use std::fmt::Display;
@@ -74,7 +74,8 @@ fn build_combobox(combo: &mut nwg::ComboBox<&'static str>, window: &nwg::Window)
 #[derive(Default)]
 pub struct ComboBox<D: Display+Default> {
     pub handle: ControlHandle,
-    collection: RefCell<Vec<D>>
+    collection: RefCell<Vec<D>>,
+    handler0: RefCell<Option<RawEventHandler>>,
 }
 
 impl<D: Display+Default> ComboBox<D> {
@@ -90,6 +91,7 @@ impl<D: Display+Default> ComboBox<D> {
             font: None,
             collection: None,
             selected_index: None,
+            v_align: VTextAlign::Center,
             parent: None
         }
     }
@@ -404,10 +406,122 @@ impl<D: Display+Default> ComboBox<D> {
         wh::send_message(handle, CB_RESETCONTENT, 0, 0);
     }
 
+    fn hook_non_client_size(&self, bg: Option<[u8; 3]>, v_align: VTextAlign) {
+        use crate::bind_raw_event_handler_inner;
+        use winapi::shared::windef::{HGDIOBJ, RECT, HBRUSH, POINT};
+        use winapi::um::winuser::{WM_NCCALCSIZE, WM_NCPAINT, WM_SIZE, DT_CALCRECT, DT_LEFT, NCCALCSIZE_PARAMS, COLOR_WINDOW};
+        use winapi::um::winuser::{SWP_NOOWNERZORDER, SWP_NOSIZE, SWP_NOMOVE, SWP_FRAMECHANGED};
+        use winapi::um::winuser::{GetDC, DrawTextW, ReleaseDC, GetClientRect, GetWindowRect, FillRect, ScreenToClient, SetWindowPos, GetWindowTextW, GetWindowTextLengthW};
+        use winapi::um::wingdi::{SelectObject, CreateSolidBrush, RGB};
+        use std::ptr;
+
+        if self.handle.blank() { panic!(NOT_BOUND); }
+        let handle = self.handle.hwnd().expect(BAD_HANDLE);
+
+        let brush = match bg {
+            Some(c) => unsafe { CreateSolidBrush(RGB(c[0], c[1], c[2])) },
+            None => COLOR_WINDOW as HBRUSH
+        };
+
+        unsafe {
+
+        let handler0 = bind_raw_event_handler_inner(&self.handle, 0, move |hwnd, msg, w, l| {
+            match msg {
+                WM_NCCALCSIZE  => {
+                    if w == 0 { return None }
+
+                    // Calculate client area height needed for a font
+                    let font_handle = wh::get_window_font(hwnd);
+                    let mut r: RECT = mem::zeroed();
+                    let dc = GetDC(hwnd);
+                    
+                    let old = SelectObject(dc, font_handle as HGDIOBJ);
+
+                    let calc: [u16;2] = [75, 121];
+                    DrawTextW(dc, calc.as_ptr(), 2, &mut r, DT_CALCRECT | DT_LEFT);
+
+                    let client_height = r.bottom - 5; // 5 is the combobox padding
+
+                    SelectObject(dc, old);
+                    ReleaseDC(hwnd, dc);
+
+                    // Calculate NC area to center text.
+                    let mut client: RECT = mem::zeroed();
+                    let mut window: RECT = mem::zeroed();
+                    GetClientRect(hwnd, &mut client);
+                    GetWindowRect(hwnd, &mut window);
+
+                    let window_height = window.bottom - window.top;
+                    let info_ptr: *mut NCCALCSIZE_PARAMS = l as *mut NCCALCSIZE_PARAMS;
+                    let info = &mut *info_ptr;
+                    match v_align {
+                        VTextAlign::Top => {
+                            info.rgrc[0].bottom -= window_height - client_height;
+                        },
+                        VTextAlign::Center => {
+                            let center = ((window_height - client_height) / 2) - 1;
+                            info.rgrc[0].top += center;
+                            info.rgrc[0].bottom -= center;
+                        },
+                        VTextAlign::Bottom => {
+                            info.rgrc[0].top += window_height - client_height;
+                        },
+                    }
+                },
+                WM_NCPAINT  => {
+                    let mut window: RECT = mem::zeroed();
+                    let mut client: RECT = mem::zeroed();
+                    GetWindowRect(hwnd, &mut window);
+                    GetClientRect(hwnd, &mut client);
+
+                    let mut pt1 = POINT {x: window.left, y: window.top};
+                    ScreenToClient(hwnd, &mut pt1);
+
+                    let mut pt2 = POINT {x: window.right, y: window.bottom};
+                    ScreenToClient(hwnd, &mut pt2);
+
+                    let top = RECT {
+                        left: 0,
+                        top: pt1.y,
+                        right: client.right,
+                        bottom: client.top
+                    };
+
+                    let bottom = RECT {
+                        left: 0,
+                        top: client.bottom,
+                        right: client.right,
+                        bottom: pt2.y
+                    };
+
+                    let dc = GetDC(hwnd);
+                    FillRect(dc, &top, brush);
+                    FillRect(dc, &bottom, brush);
+                    ReleaseDC(hwnd, dc);
+                },
+                WM_SIZE => {
+                    SetWindowPos(hwnd, ptr::null_mut(), 0, 0, 0, 0, SWP_NOOWNERZORDER | SWP_NOSIZE | SWP_NOMOVE | SWP_FRAMECHANGED);
+                },
+                _ => {}
+            }
+
+            None
+        });
+
+        *self.handler0.borrow_mut() = Some(handler0.unwrap());
+
+        }
+    }
+
 }
 
 impl<D: Display+Default> Drop for ComboBox<D> {
     fn drop(&mut self) {
+        let handler = self.handler0.borrow();
+        if let Some(h) = handler.as_ref() {
+            drop(unbind_raw_event_handler(h));
+        }
+
         self.handle.destroy();
     }
 }
@@ -422,6 +536,7 @@ pub struct ComboBoxBuilder<'a, D: Display+Default> {
     font: Option<&'a Font>,
     collection: Option<Vec<D>>,
     selected_index: Option<usize>,
+    v_align: VTextAlign,
     parent: Option<ControlHandle>
 }
 
@@ -477,6 +592,10 @@ impl<'a, D: Display+Default> ComboBoxBuilder<'a, D> {
         self
     }
 
+    pub fn v_align(mut self, align: VTextAlign) -> ComboBoxBuilder<'a, D> {
+        self.v_align = align;
+        self
+    }
 
     pub fn build(self, out: &mut ComboBox<D>) -> Result<(), NwgError> {
         let flags = self.flags.map(|f| f.bits()).unwrap_or(out.flags());
@@ -518,6 +637,8 @@ impl<'a, D: Display+Default> ComboBoxBuilder<'a, D> {
         if self.focus {
             out.set_focus();
         }
+
+        out.hook_non_client_size(None, self.v_align);
 
         Ok(())
     }