+++ changelog.md
@@ -1,6 +1,7 @@
 1.0.9
 * BREAKING CHANGE: File dialog `get_selected_item` & `get_selected_items` now return `OsString` instead of String to handle some exotic Windows path
-* Added vertical alignment to label
+* Added vertical alignment  `v_align` to label (defaults to center)
+* Added vertical alignment `v_align` to combobox (defaults to center)
 * Fixed multi-line label alignment 
 * Added helpers functions to Bitmap and Icons
 * Added the `Monitor` struct to query monitor and screen information